requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==1.0.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.4
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.0
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
blessed==1.20.0
blinker==1.7.0
boto3==1.29.6
botocore==1.32.6
Brotli==1.1.0
CacheControl==0.14.0
cachelib==0.9.0
cachetools==5.3.3
celery==5.4.0
cement==2.10.14
certifi==2024.8.30
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
comm==0.1.4
constantly==23.10.4
contourpy==1.1.1
cryptocode==0.1
cryptography==41.0.7
cssselect2==0.7.0
cycler==0.12.1
daphne==4.0.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.8.0rc2
distro==1.9.0
Django==5.1.3
django-allauth==65.2.0
django-appconf==1.1.0
django-channels-jwt-auth-middleware==1.0.0
django-compressor==4.4
django-cors-headers==4.3.0
django-dbbackup==4.2.1
django-extensions==3.2.3
django-mathfilters==1.0.0
django-migration-linter==5.1.0
django-pwa==1.1.0
django-ratelimit==3.0.1
django-recaptcha==4.0.0
django-redis==5.4.0
django-storages==1.14.2
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
docutils==0.21.2
drf-yasg==1.21.7
ecdsa==0.19.0
et-xmlfile==1.1.0
executing==2.0.0
exponent-server-sdk==2.0.0
Faker==21.0.0
filelock==3.17.0
firebase-admin==6.5.0
flake8==6.1.0
Flask==3.0.0
Flask-Caching==2.3.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.43.1
frozenlist==1.4.1
fsspec==2024.12.0
ghp-import==2.1.0
git-filter-repo==2.45.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
hyperlink==21.0.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.16.1
itsdangerous==2.1.2
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.1
Jinja2==3.1.2
jiter==0.5.0
jmespath==1.0.1
jupyter_client==8.5.0
jupyter_core==5.4.0
kiwisolver==1.4.5
kombu==5.3.7
mailjet-rest==1.3.4
Markdown==3.7
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mediapipe==0.10.14
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-pdf-export-plugin==0.5.10
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
nest-asyncio==1.5.8
networkx==3.4.2
numpy==1.26.1
oauthlib==3.2.2
openai==1.47.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==24.2
pandas==1.5.3
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
pipdeptree==2.23.1
platformdirs==3.11.0
pluggy==1.3.0
pockets==0.9.1
prompt-toolkit==3.0.39
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.6
psycopg==3.1.17
psycopg-binary==3.1.17
psycopg2==2.9.9
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydotplus==2.0.2
pydyf==0.11.0
pyflakes==3.1.0
Pygments==2.19.1
PyJWT==2.8.0
PyMySQL==1.0.2
pyOpenSSL==23.3.0
pyparsing==3.1.1
pyphen==0.16.0
pyshorteners==1.0.1
PySocks==1.7.1
pyspark==3.5.3
pytest==7.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==25.1.1
rcssmin==1.1.1
redis==5.1.0
render-python==2.3.1
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==1.3.1
rjsmin==1.2.1
rsa==4.7.2
rst2pdf==0.102
s3transfer==0.7.0
schedule==1.2.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.21.0
semantic-version==2.10.0
service-identity==23.1.0
six==1.16.0
smartypants==2.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
sounddevice==0.5.0
soupsieve==2.5
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.35
sqlparse==0.5.2
stack-data==0.6.3
stripe==11.4.1
sympy==1.13.3
termcolor==2.4.0
tinycss2==1.3.0
toml==0.10.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.3.3
tqdm==4.66.5
traitlets==5.12.0
trio==0.25.1
trio-websocket==0.11.1
tweepy==4.14.0
twilio==9.2.3
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
watchdog==5.0.2
wcwidth==0.2.8
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.8.2
wsproto==1.2.0
yarl==1.9.4
zipp==3.20.2
zope.interface==6.1
zopfli==0.2.3
